+++ CHANGELOG.md
@@ -1,9 +1,15 @@
+# 0.9.57 Release:
+## Bug Fixes:
+- Favored Foe
+- Slasher
+## Update Notes:
+- Manual Rolling has been updated. Each user can now be individually configured.
+- Support for D&D system version 3.0.0 will be in a future update. Do not update your system yet. This setting will still only effect attacks, saves, and damage rolls from midi-qol workflows.
+  
 # 0.9.56 Release:
 ## Bug Fixes:
 - Manifest Echo
 - Item configuration
-## Update Notes:
-- Support for D&D system version 3.0.0 will be in a future update. Do not update your system yet.
   
 # 0.9.55 Release:
 ## Bug Fixes: